abstracted out some methods & made width settable
