+++ CN1CircleProgress/src/com/codename1/progress/CircleProgress.java
@@ -35,7 +35,7 @@ import com.codename1.ui.plaf.Style;
  *
  * @author Chen
  */
-public class CircleProgress extends Slider {
+public class CircleProgress extends BaseRoundProgress {
 
     private int arcWidth = 10;
     
@@ -119,21 +119,5 @@ public class CircleProgress extends Slider {
         }
     }
 
-    public void paintComponentBackground(Graphics g) {
-    }
-
-    @Override
-    protected Dimension calcPreferredSize() {
-        Style style = getStyle();
-        int prefW = Display.getInstance().getDisplayWidth()*30/100, prefH = prefW;
-        
-        prefH += (style.getPadding(false, Component.TOP) + style.getPadding(false, Component.BOTTOM));
-        prefW += (style.getPadding(isRTL(), Component.RIGHT) + style.getPadding(isRTL(), Component.LEFT));
-        
-        return new Dimension(prefW, prefH);
-    }
-    
-    
-    
 
 }
